refactor(cf): relocate cf-apps signal config out of list framework
