+++ src/frontend/packages/cloud-foundry/src/shared/signal-list-configs/app/cf-apps-signal-config.service.ts
@@ -3,16 +3,16 @@ import { toSignal } from '@angular/core/rxjs-interop';
 import { HttpClient, HttpParams } from '@angular/common/http';
 import { firstValueFrom } from 'rxjs';
 import type { EndpointModel } from '@stratosui/store';
-import { CnsiAppsSource } from '../../../../../services/data-sources/cnsi-apps-source';
-import { CnsiRoutesSource } from '../../../../../services/data-sources/cnsi-routes-source';
-import { CnsiServiceBindingsSource } from '../../../../../services/data-sources/cnsi-service-bindings-source';
-import { MergeOrchestrator } from '../../../../../services/data-sources/merge-orchestrator';
-import { EndpointDataRegistry } from '../../../../../services/endpoint-data/endpoint-data.registry';
-import { ViewPipeline, SortSpec } from '../../../../../services/data-sources/view-pipeline';
-import type { StApp, StAppRoutesResponse, StOrg, StOrgsResponse, StRoute, StServiceCredentialBinding, StServiceCredentialBindingsResponse, StSpace, StSpacesResponse } from '../../../../../services/endpoint-data/stratos-types';
-import { CloudFoundryService } from '../../../../data-services/cloud-foundry.service';
-import { writeWithJob } from '../../../../../services/async-jobs/write-with-job';
-import type { StratosJob } from '../../../../../services/async-jobs/async-job.types';
+import { CnsiAppsSource } from '../../../services/data-sources/cnsi-apps-source';
+import { CnsiRoutesSource } from '../../../services/data-sources/cnsi-routes-source';
+import { CnsiServiceBindingsSource } from '../../../services/data-sources/cnsi-service-bindings-source';
+import { MergeOrchestrator } from '../../../services/data-sources/merge-orchestrator';
+import { EndpointDataRegistry } from '../../../services/endpoint-data/endpoint-data.registry';
+import { ViewPipeline, SortSpec } from '../../../services/data-sources/view-pipeline';
+import type { StApp, StAppRoutesResponse, StOrg, StOrgsResponse, StRoute, StServiceCredentialBinding, StServiceCredentialBindingsResponse, StSpace, StSpacesResponse } from '../../../services/endpoint-data/stratos-types';
+import { CloudFoundryService } from '../../data-services/cloud-foundry.service';
+import { writeWithJob } from '../../../services/async-jobs/write-with-job';
+import type { StratosJob } from '../../../services/async-jobs/async-job.types';
 import type { SignalListDropdownOption } from '@stratosui/core';
 import { ListStateStore, naturalCompare } from '@stratosui/core';
